refactor: split manifest into own file, extract shared priority validation
- Move Manifest() (200+ lines of JSON schemas) from linear.go to
  manifest.go — reduces linear.go from 446 to ~230 lines
- Extract validatePriority() shared by create_issue and update_issue,
  eliminating duplicated range-check logic

No behavior changes. All tests pass.
