+++ connectors/linear/update_issue.go
@@ -27,8 +27,8 @@ func (p *updateIssueParams) validate() error {
 	if p.IssueID == "" {
 		return &connectors.ValidationError{Message: "missing required parameter: issue_id"}
 	}
-	if p.Priority != nil && (*p.Priority < 0 || *p.Priority > 4) {
-		return &connectors.ValidationError{Message: "priority must be 0 (none), 1 (urgent), 2 (high), 3 (medium), or 4 (low)"}
+	if err := validatePriority(p.Priority); err != nil {
+		return err
 	}
 	if !p.hasUpdates() {
 		return &connectors.ValidationError{Message: "at least one field to update must be provided (title, description, assignee_id, priority, state_id, or label_ids)"}
